feat(-): pages 폴더 구조 수정, type import 옵션 수정
