+++ src/pages/LoginPage/index.tsx
@@ -1,5 +1,6 @@
-import { View, Text, Button, StyleSheet } from 'react-native';
-import useLogin from './useLogin';
+import { View, Text, Button } from 'react-native';
+import useLogin from './hooks/useLogin';
+import styles from './styles';
 
 const LoginPage = () => {
   const { handleLogin } = useLogin();
@@ -11,21 +12,5 @@ const LoginPage = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: 'center',
-    backgroundColor: '#f7f7f7',
-    flex: 1,
-    justifyContent: 'center',
-    padding: 16,
-  },
-  title: {
-    color: '#333',
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-  },
-});
 
 export default LoginPage;
